Add additional location for coastal properties
The purpose of this change is to allow the basic coastal.properties file
to specify default values for properties (so that they do not have to be
included in every property file) but -- at the same time -- to allow
users to override these defaults.

Therefore the Configuration is now loaded from three places:

1. "coastal.properties" in the built-in resources
2. ".coastal/coastal.properties" in the user's home directory
3. the filename specified on the command line.

If the second location is not found, it is ignored.