Add coverage for conditions
As expected all conditions are covered completely, as far as that is
possible.  This is also true for instruction coverage.

The purpose of these listeners is not so much to compute the final
coverage, but to monitor the progress of the coverage and switch
strategies.  Further modification of the coverage listeners are required
so that coverage can be queried on-the-fly.  Moreover, it is also not
clear at the moment how to connect such a meta-strategy to the listener.