Added latest green library back in
It was temporarily removed while green modifications were in progress.

More detail:  to modify green to support new COASTAL features or just to
debug a potential bug, the gradle dependency on green is removed.  In
Eclipse, the green project is added to the build path.

Once the bug/feature has been addressed, the process is reversed.
Except that we sometimes forget.  The correct order should be:
short-circuit green in COASTAL, fix green, commit and publish green,
un-short-circuit green in COASTAL.