Further fixes moving closer to working refactored code
The refactoring of `InstrumentationClassLoader` brought a serious issue
to light: the class loader should be re-created for each run, since
"leftover" data from a previous run may just interfere with the current
run.

This would mean that many more class reloads and many more
instrumentations are required (although the latter could be cached).
Unfortunately, the matter is not so straightforward: apart from the
speed penalty, there is also the question of whether the system classes
*could* be reloaded in this way.