Major refactor to remove non-static instances
This is a big change.  But it is necessary to accommodate an important
feature: the ability to handle static data.

Firstly, the classloader needs to be re-started to allow classes to
reload.  The actual instrumentation is cached in the new
`InstrumentationClassManager`.

Secondly, `SymbolicState` needs to be an instance (and not static) so
that static data will not be carried over from one run to the next.

The refactoring also revealed a problem with testing:  Eclipse and
Gradle use different compilers and therefore the same class can be
instrumented differently.  So far, it seems to be limited to line
numbers, but the bottom line is that the gradle tests are definitive and
the in-Eclipse unit testing is (or can be) inaccurate.  The testing
framework was made more robust, but it still cannot guarantee that the
two compilers agree.