Improve handling of actual return values
Before, the return values from non-instrumented functions were treated
as symbolic, but beyond the control of COASTAL.  Then the default
changed so that the actual return value was used.  This broke the basic
assumption that every path contains the same conjuncts (except when
dealing with switch statements which are handled differently).

Willem's suggested an alternative: use the symbolic values, but add a
passive ("on-the-side") constraint that the symbolic variable has a
specific value.

This is implemented in this change.