Fix several small and one big issue to make Mystery work [ci skip]
The big issue was that all delegates were broken.  Then SymbolicVM state
class was not "connected" properly to the specific SymbolicState that
was associated with a Diver thread.  This is because the delegates are
not loaded by the instrumenting classloaders but instead by the
classloader of the main thread.  This was not an issue before, but
multi-threading requires that we pass the specific symbolic state to the
delegate methods.