Add property files for static examples [ci skip]
These files will be moved to the test section soon.  But first we have
to figure out how to make them into sensible tests that could fail if
the static support is not adequate.

Also fixed a bug in the trace-all mode that did not remove all the
operands of a IF_ICMP*.

Also tried to fix Travis CI gradle version.