Introduce adapter interface between static VM and non-static state
SymbolicVM.java has been renamed to VM.java, and now uses State.java
which is an interface.  This is extended by SymbolicState.java.  In
other words, State.java has been introduced as an abstraction for
SymbolicState.  This is in preparation for the next step which
introduces a second kind of symbolic state for traces.