Remove checks for duplicate/constant conjuncts [ci skip]
Duplicate and constant conjuncts were required when the path tree
checked the identity of branch conditions.  With the introduction of
support for switch branches, this requirement fell away.  Now it is more
convenient to "ignore" constant and duplicate conjuncts and to tree them
as "normal" conjuncts.

We can now reconcile the path tree constructed by segmented path
condition and by traces.

This closes #17.