Important changes to make fuzzing work [ci skip]
The most important:

- Surfers (not yet divers) abort the run of the program when the
symbolic exploration is complete, instead of carrying on with the
non-symbolic execution until the end of the program.

- Surfers (not yet divers) will abort the run as soon as they detect
that the execution path has been fully explored.

- Of the various fuzzers, Cybrid seems to work best.  It is very
different from other fuzzing strategies and may have to be adjusted in
the future.

- Traces can carry arbitrary payloads.

- All IF instructions now path the operands on the stack on to the
symbolic execution engine.  Surfers use this information to detect the
outcome of the IF, instead of the post-IF check that divers still use.
This will change in the near future.