First pass at implementing ANEWARRAY
Support for general arrays is shaky.  Not tested in a wide enough range
of situations.  Coming soon.  Closes #29 for now (but we might need to
reopen it).

In particular, the instructions as implemented handle the easy case
where array sizes and indices are constants.  A more general
implementation (for variable size & indices) is problably needed but may
need us to switch our solver.