Allow arrays of tasks instead of single tasks
Sometimes, a task factory wants to create not just a single thread for
each instance of that task, but an array of threads.  The terminology
sucks, because Task is the base class for such threads, but also the
collective name for a set of threads that corresponds to one instance of
the task.  Might fix the names later on.