Reset to master version
In this new branch something is broken.  But I do not know what.  The
plan is to reset as much of the new branch to the master version, and
then introduce the changes slowly back into the master.

In this file, it is just one configuration statement, but it is
important to have as few differences as possible.