Fix bug in stack entry size handling
When a local variable is loaded onto the stack, char/byte/short should
be converted to int's.  This was not done and produced problems when
constraints were generated for the solver.

This was fixed by not correcting the way the stack works, but rather the
way constraints are output to the solver.  This is bad, because the
symbolic state is now still inaccurate (although it is never 100%
accurate), but it is good because it is more robust.